worker/requirements.txt
python-dotenv
redis
channels
